+++ examples/src/passes/guassian/index.ts
@@ -1,27 +1,39 @@
 import {
-  fetchImage,
+  CopyPass,
+  DataType,
+  fetchHDR,
+  InternalFormat,
   makeBufferGeometryFromGeometry,
   makeProgramFromShaderMaterial,
-  Orbit,
   passGeometry,
+  PixelFormat,
   renderBufferGeometry,
   RenderingContext,
   ShaderMaterial,
   TexImage2D,
   Texture,
-  TextureFilter,
-  textureToTexImage2D,
-  TextureWrap
+  TextureEncoding,
+  textureToTexImage2D
 } from '@threeify/core';
-import {
-  Mat4,
-  mat4Inverse,
-  mat4PerspectiveFov,
-  quatToMat4
-} from '@threeify/math';
 
+import { getThreeJSHDRIUrl, ThreeJSHRDI } from '../../utilities/threejsHDRIs';
 import fragmentSource from './fragment.glsl';
 import vertexSource from './vertex.glsl';
+
+let blurRadius = 16;
+
+document.addEventListener('keydown', (event) => {
+  switch (event.key) {
+    case 'ArrowUp':
+      blurRadius = Math.min(blurRadius * 2, 128);
+      break;
+    case 'ArrowDown':
+      blurRadius = Math.max(blurRadius / 2, 1);
+      break;
+  }
+
+  console.log('blurRadius', blurRadius);
+});
 
 async function init(): Promise<void> {
   const geometry = passGeometry();
@@ -35,60 +47,51 @@ async function init(): Promise<void> {
     document.getElementById('framebuffer') as HTMLCanvasElement
   );
 
-  let imageIndex = 0;
-
-  const images = [];
-  const textures: Texture[] = [];
-  const texImage2Ds: TexImage2D[] = [];
-  for (let i = 0; i < 5; i++) {
-    images.push(
-      fetchImage(`/assets/textures/cube/kitchen/kitchenb_${i + 1}.jpg`).then(
-        (image) => {
-          const texture = new Texture(image);
-          texture.wrapS = TextureWrap.ClampToEdge;
-          texture.wrapT = TextureWrap.ClampToEdge;
-          texture.minFilter = TextureFilter.Linear;
-          textures[i] = texture;
-          texImage2Ds[i] = textureToTexImage2D(context, texture);
-        }
-      )
-    );
-  }
-
-  await Promise.all(images);
+  const hdrPromise = fetchHDR(
+    getThreeJSHDRIUrl(ThreeJSHRDI.royal_esplanade_1k)
+  );
+  const latLongTexturePromise = hdrPromise.then((hdr) => {
+    return new Texture(hdr);
+  });
 
   const { canvasFramebuffer } = context;
   window.addEventListener('resize', () => canvasFramebuffer.resize());
 
-  const orbit = new Orbit(context.canvas);
+  const texture = await latLongTexturePromise;
+  const rgbeTexImage2D = textureToTexImage2D(context, texture);
+  const hdrTexImage2D = new TexImage2D(
+    context,
+    [rgbeTexImage2D.size],
+    InternalFormat.RGBA16F,
+    DataType.HalfFloat,
+    PixelFormat.RGBA
+  );
+
+  const copyPass = new CopyPass(context);
+  await copyPass.exec({
+    sourceTexImage2D: rgbeTexImage2D,
+    sourceEncoding: TextureEncoding.RGBE,
+    targetTexImage2D: hdrTexImage2D,
+    targetEncoding: TextureEncoding.Linear
+  });
+
+  hdrTexImage2D.generateMipmaps();
+  rgbeTexImage2D.dispose();
 
   const passProgram = await makeProgramFromShaderMaterial(
     context,
     passMaterial
   );
   const passUniforms = {
-    viewToWorld: new Mat4(),
-    screenToView: mat4Inverse(
-      mat4PerspectiveFov(30, 0.1, 4, 1, canvasFramebuffer.aspectRatio)
-    ),
-    equirectangularMap: texImage2Ds[0]
+    standardDeviation: blurRadius,
+    texture: hdrTexImage2D
   };
   const bufferGeometry = makeBufferGeometryFromGeometry(context, geometry);
 
   function animate(): void {
     requestAnimationFrame(animate);
 
-    passUniforms.viewToWorld = mat4Inverse(quatToMat4(orbit.rotation));
-    passUniforms.screenToView = mat4Inverse(
-      mat4PerspectiveFov(
-        15 * (1 - orbit.zoom) + 15,
-        0.1,
-        4,
-        1,
-        canvasFramebuffer.aspectRatio
-      )
-    );
-    passUniforms.equirectangularMap = texImage2Ds[imageIndex];
+    passUniforms.standardDeviation = blurRadius;
 
     renderBufferGeometry({
       framebuffer: canvasFramebuffer,
@@ -96,46 +99,9 @@ async function init(): Promise<void> {
       uniforms: passUniforms,
       bufferGeometry
     });
-
-    orbit.update();
   }
 
   animate();
-
-  window.addEventListener(
-    'keydown',
-    (event) => {
-      if (event.key !== undefined) {
-        imageIndex =
-          ((event.key.codePointAt(0) || 0) - ('0'.codePointAt(0) || 0)) %
-          images.length;
-        // Handle the event with KeyboardEvent.key and set handled true.
-      }
-    },
-    true
-  );
 }
 
 init();
-
-window.addEventListener(
-  'keydown',
-  (event) => {
-    if (event.defaultPrevented) {
-      return; // Should do nothing if the default action has been cancelled
-    }
-
-    const handled = false;
-    if (event.key !== undefined) {
-      // Handle the event with KeyboardEvent.key and set handled true.
-    } else if (event.keyCode !== undefined) {
-      // Handle the event with KeyboardEvent.keyCode and set handled true.
-    }
-
-    if (handled) {
-      // Suppress "double action" if event handled
-      event.preventDefault();
-    }
-  },
-  true
-);